Bundling with tsup, extracting schema.
