--- build.ts
@@ -1,11 +1,0 @@
-await Bun.build({
-  entrypoints: ["src/index.ts"],
-  outdir: "dist",
-  /** @link https://github.com/oven-sh/bun/issues/159 */
-  format: "esm",
-  splitting: false,
-  sourcemap: "none",
-  minify: true,
-  target: "node",
-  external: ["*"],
-});
