+++ package.json
@@ -26,8 +26,7 @@
     "*.md"
   ],
   "scripts": {
-    "build:cjs": "tsc -p tsconfig.cjs.json && uglifyjs dist/index.js -c -m -o dist/index.cjs",
-    "build": "bun run build:cjs && bun run build.ts",
+    "build": "tsup",
     "lint": "bun run biome check",
     "test": "tsc --noEmit && bun test src",
     "mdfix": "bunx --bun prettier *.md --write",
@@ -45,8 +44,8 @@
     "@types/ramda": "^0.30.1",
     "@typescript-eslint/rule-tester": "^8.0.1",
     "json-schema-to-ts": "^3.1.0",
-    "typescript": "^5.5.4",
-    "uglify-js": "^3.19.1"
+    "tsup": "^8.2.4",
+    "typescript": "^5.5.4"
   },
   "peerDependencies": {
     "eslint": "^9.0.0",
