--- tsconfig.cjs.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@tsconfig/node18/tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "module": "commonjs",
-    "moduleResolution": "node",
-  },
-  "include": ["src/index.ts"]
-}
